DETR-Insulator-Detection/requirements.txt
# PyTorch with CUDA 12.1 support (for Lightning.ai L4 GPU)
torch>=2.0.0
torchvision>=0.15.0

# Transformers and DETR dependencies
transformers>=4.30.0

# Image processing
Pillow>=10.0.0

# Core ML libraries
numpy>=1.24.0
scikit-learn>=1.3.0

# COCO evaluation tools
pycocotools>=2.0.6

# Configuration management
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0

# Roboflow for dataset download
roboflow>=1.1.0

# Ultralytics YOLO (optional, for HITL comparison)
ultralytics>=8.0.0
